+++ Vireo_VS/VireoCommandLine.vcxproj.filters
@@ -232,5 +232,8 @@
     <ClInclude Include="..\source\include\Array.h">
       <Filter>VireoSource\Include</Filter>
     </ClInclude>
+    <ClInclude Include="..\source\include\DataReflectionVisitor.h">
+      <Filter>VireoSource\Include</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
